x210: Fix up battery reporting information
The EC appears to be setting CAPCITY_MODE in the battery to 1, which means
the battery reports in centiWatts and centiWatt hours. However, the EC is
also retrieving the current in mA and not doing any conversion. It's
impossible to represent this correctly in ACPI, so we need to fix it up.
The easiest way to do so is to report in mW, multiply the raw battery values
by 10, and then multiply the current by the voltage in order to get a more
accurate live consumption rate.

Change-Id: Id90a38ee143b37dacf1705cfd2f37ca09dfc138f